Refactor `Schools::Wizardable` concern
Before, this concern could only be used for changing ECT details.

This refactors the concern so we can define a wizardable type (e.g `ect` or
`mentor`).

Now, the concern can be used in ECT and Mentor controllers.
