+++ app/controllers/schools/ects/change_email_address_wizard_controller.rb
@@ -3,6 +3,8 @@ module Schools
     class ChangeEmailAddressWizardController < SchoolsController
       include Wizardable
 
+      wizard_for :ect
+
       def new
         render @current_step
       end
