+++ app/controllers/schools/ects/change_name_wizard_controller.rb
@@ -3,6 +3,8 @@ module Schools
     class ChangeNameWizardController < SchoolsController
       include Wizardable
 
+      wizard_for :ect
+
       def new
         render @current_step
       end
