+++ app/controllers/schools/ects/change_working_pattern_wizard_controller.rb
@@ -3,6 +3,8 @@ module Schools
     class ChangeWorkingPatternWizardController < SchoolsController
       include Wizardable
 
+      wizard_for :ect
+
       def new
         render @current_step
       end
